+++ glogarch/__init__.py
@@ -1,3 +1,3 @@
 """Glogarch - Graylog Open Archive tool."""
 
-__version__ = "1.7.9"
+__version__ = "1.7.10"